pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "mcp-florence2"
version = "0.1.0"
description = "An MCP server for processing images using Florence-2"
readme = "README.md"
authors = [
    { name = "Junpei Kawamoto", email = "kawamoto.junpei@gmail.com" },
]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "click>=8.1.8",
    "dill>=0.3.9",
    "einops>=0.8.1",
    "mcp>=1.2.1",
    "pillow>=11.1",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "timm>=1.0.14",
    "torch>=2.6",
    "transformers>=4.49",
]
scripts.mcp-florence2 = "mcp_florence2:main"

[dependency-groups]
dev = [
    "pre-commit>=4.1",
    "pre-commit-uv>=4.1.4",
    "pytest>=8.3.4",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pyproject-fmt]
indent = 4

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
